Support --arch and --os flags for arkade get
By default, the `arkade get` command will use the client
cpu architecture and os (using `uname` command).
The new flags will allow to override the default behavior, and download
different binaries for different platforms.

In case of an unsupported architecture or OS, an appropriate message
would be shown with all the options available.

Signed-off-by: Yuval Goldberg <yuvigoldi@gmail.com>